Update SSL-checking for Let's Encrypt.
Now that I'm using Let's encrypt I have a lot of SSL-certificates
to check - so this update changes the `ssl-expiry-date` test to
allow testing multiple domains at once.

Because we've switched from:

        ssl-expiry-date $host $port

To:

        ssl-expiry-date $host1 $host2  .. $hostN

We've now had to add getopt-based parsing to specify a different
port, which is achived via "`-p 1234`".  We've also added a simpler
output which is just "`$domain: $days`".